requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
neo4j
python-dotenv
pydantic
pydantic-settings
